requirements.txt
# Install with: `pip install -r requirements.txt`

# Infrastructure
Django == 1.8
djangorestframework==3.3.3

